requirements.txt
pydantic>=2.0
google-genai>=1.0
fastapi>=0.110
uvicorn>=0.29
websockets>=12.0
pytest>=8.0
numpy>=1.26
